Moved more logic to the abstract class.
